pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "KrishiSahayak"
version = "1.0.0"
authors = [
  { name="Vikas Sahani", email="vikassahani17@gmail.com" },
]
description = "AI-Powered Crop Health Assistant for farmers."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
# Core dependencies required for the project to be imported
# UPDATED versions for security, performance, and compatibility
dependencies = [
    "numpy>=1.26",
    "Pillow>=10.0",
    "PyYAML>=6.0",
    "torch>=2.3",
]

[project.urls]
Homepage = "https://github.com/VIKAS9793/KrishiSahayak"
Repository = "https://github.com/VIKAS9793/KrishiSahayak"
"Bug Tracker" = "https://github.com/VIKAS9793/KrishiSahayak/issues"

[project.optional-dependencies]
# Dependencies for training the model
train = [
    "KrishiSahayak[data]",
    # UPDATED versions
    "pytorch-lightning>=2.2",
    "torchvision>=0.18",
    "timm>=0.9.12",
    "albumentations>=1.4",
    "torchmetrics>=1.4",
]
# Dependencies for data preparation and validation
data = [
    # UPDATED versions
    "pandas>=2.2",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "opencv-python-headless",
    "scikit-learn>=1.4",
    "tqdm",
]
# Dependencies for deployment and inference
deploy = [
    # UPDATED versions
    "onnx>=1.16",
    "onnxruntime>=1.18",
    "tensorflow>=2.16",
    "onnx-tf>=1.10",
]
# Dependencies for testing
test = [
    "pytest>=7.4",
]
# Dependencies for development (linting, formatting, etc.)
dev = [
    "KrishiSahayak[train,test]",
    "black>=24.0",
    "ruff>=0.4",
    "mypy>=1.9",
    "pre-commit",
    "pip-tools", # Added for generating locked requirements files
]
# A group to install everything
all = [
    "KrishiSahayak[dev,deploy]",
]

# --- Tool Configurations ---

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff]
line-length = 88
select = ["E", "W", "F", "I", "UP", "C4"] # Standard lints + import sorting
ignore = ["E501"] # Ignore line too long, as black handles it

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
ignore_missing_imports = true